add example

+++ include/graphics/Mesh.hpp
@@ -9,12 +9,12 @@ namespace CpuRasterizor
 	{
 	public:
 		std::vector<Vertex> vertices;
-		std::vector<uint32_t> indices;
+		std::vector<size_t> indices;
 
 	public:
 		Mesh();
 		Mesh(const Mesh& other);
-		Mesh(const std::vector<Vertex>& _vertices, const std::vector<uint32_t>& _indices);
+		Mesh(const std::vector<Vertex>& _vertices, const std::vector<size_t>& _indices);
 		~Mesh();
 		Mesh& operator= (const Mesh& other);
 	};
